fix(Interactions): split properly into base and sub-structure
